requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0.2
django-cors-headers==4.9.0
django-oauth-toolkit==3.2.0
djangorestframework==3.16.1
idna==3.11
jwcrypto==1.5.6
oauthlib==3.3.1
pillow==12.1.0
pycparser==3.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
